+++ easy_setting/CHANGELOG.md
@@ -1,3 +1,6 @@
+## 0.0.4
+* Refactor. Move from Firestore to Realtime database.
+
 ## 0.0.3
 * Update. easy_firestore.